+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new command `!lastonline <char>` as non-org equivalent for `!lastseen`
 - Add new setting `orglist_pork_url` to allow downloading orglists from our reverse proxy. This solves PORK being down very often lately.
+- Add new settings to allow automatically sending random greetings to players joining the private channel.
 
 ### Changed